change video datetime field to date and order_in_day
