+++ meetingsvideos/templates/meetingsvideos/meeting_detail.html
@@ -6,6 +6,10 @@
 
 {% for video in meeting.videos_by_time %}
 
+{% if video.order_in_day == 1 %}
+    <h2>{{ video.date }}</h2>
+{% endif %}
+
 {% include "meetingsvideos/video_preview_symposium.html" %}
 
 {% endfor %}
