+++ uploadvideos.py
@@ -13,28 +13,16 @@ from meetingsvideos.models import (
 
 
 # converts EDTF date to datetime object
-def process_time(str):
-    lst = str.split("T")
-
+def process_date(str):
     # process year, month, and day
-    ymd = lst[0].split("-")
+    ymd = str.split("-")
 
     year = int(ymd[0])
     month = int(ymd[1])
     day = int(ymd[2])
 
-    # date_obj = date(year, month, day)
+    return datetime.datetime(year, month, day, tzinfo=ZoneInfo("America/New_York"))
 
-    # process time
-    hms = lst[1].split(":")
-    hour = int(hms[0])
-    minute = int(hms[1])
-
-    # time_obj = time(hour, minute)
-
-    return datetime.datetime(
-        year, month, day, hour, minute, tzinfo=ZoneInfo("America/New_York")
-    )
 
 
 def process_diglib_url(str):
@@ -67,7 +55,7 @@ def process_affiliation(position, institution, meeting, speaker):
 
 
 # create speaker object and add to video
-# only process display name and affiliation - LCSH will be handled with other LCSH
+# only process display name and affiliation - speaker LCSH will be handled with other LCSH
 def add_speaker_to_video(video, display_name, position_1, institution_1, position_2, institution_2, meeting):
     speaker, created = Speaker.objects.get_or_create(
             display_name=display_name
@@ -82,7 +70,6 @@ def add_speaker_to_video(video, display_name, position_1, institution_1, positio
         process_affiliation(position_1, institution_1, meeting, speaker)
     if position_2 or institution_2:
         process_affiliation(position_2, institution_2, meeting, speaker)
-        print("if position_2 or institution_2 runs")
     video.speakers.add(speaker)
 
 
@@ -102,7 +89,7 @@ def process_symposium(str, meeting):
 
 
 # process individual spreadsheet row to create video
-def process_video(row):
+def process_video(row, n, prev_date):
     print("\n-----------\nVIDEO: " + row["title"] + "\n")
 
     # find correct meeting - search by name
@@ -110,9 +97,13 @@ def process_video(row):
 
     # find or create symposium
     symposium, reminders = process_symposium(row["symposium"], meeting)
-
-    # create date and time
-    time_obj = process_time(row["time"])
+    
+    # create date object
+    date=process_date(row["date"])
+    if date == prev_date:
+        n += 1
+    else:
+        n = 1
 
     # TODO: let this update video object if not all data matches? which fields should ID it?
     video, created = Video.objects.get_or_create(
@@ -128,17 +119,14 @@ def process_video(row):
         admin_category=row["admin_category"],
         meeting=meeting,
         symposium=symposium,
-        time=time_obj,
+        date=date,
+        order_in_day=n,
     )
 
     # if record for this video already exists, alert user
     if not created:
         print("Video already exists in database; no new record created")
-        return (
-            "Video "
-            + row["title"]
-            + " already exists in database; no new record created\n"
-        )
+        return n, None, None
     # if record is new, add remaining details
     else:
         video.save()
@@ -167,15 +155,6 @@ def process_video(row):
                 meeting,
             )
 
-        # get or create LCSH
-        # add_lcsh_to_video(
-        #     video,
-        #     row["lcsh_topic"],
-        #     row["lcsh_geographic"],
-        #     row["lcsh_temporal"],
-        #     row["lcsh_name_personal"],
-        #     row["lcsh_name_corporate"],
-        # )
 
         # add department and discipline
         departments = add_category_to_video(row["aps_departments"], APSDepartment)
@@ -190,7 +169,7 @@ def process_video(row):
             for discipline in disciplines:
                 video.academic_disciplines.add(discipline)
 
-        return reminders
+        return n, date, reminders
 
 
 # loop through spreadsheet, adding a video for each row
@@ -199,10 +178,15 @@ def add_videos():
 
     with open("videos.csv", newline="", encoding="utf8") as csvfile:
         reader = csv.DictReader(csvfile)
+        
+        # set up variables to determine order in day
+        n = 1
+        date = datetime.datetime(1900, 1, 1, tzinfo=ZoneInfo("America/New_York"))
 
         for row in reader:
             try:
-                reminders += process_video(row)
+                n, date, new_reminders = process_video(row, n, date)
+                reminders += new_reminders
             except Exception as e:
                 print(f"An error occurred: {e}")
                 reminders += "AN ERROR OCCURRED: " + row["title"] + "\n"
